+++ project/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './components/app/app';
 import { store } from './store';
-import { fetchData } from './store/action';
+import { fetchData } from './store/api-action';
 
 store.dispatch(fetchData());
