+++ project/src/store/api-action.ts
@@ -1,0 +1,25 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { AxiosInstance } from 'axios';
+import { Offer } from '../components/app/app-props';
+
+
+export const Action = {
+  CITY_CHANGE: 'CITY_CHANGE',
+  OFFERS_LIST_FILL: 'OFFERS_LIST_FILL',
+  SORT_BY_CHANGE_ACTION: 'SORT_BY_CHANGE_ACTION',
+  REQUIRE_AUTHORIZATION: 'REQUIRE_AUTHORIZATION',
+};
+
+const isAxiosInstance = (extra: unknown): extra is AxiosInstance => 'get' in (extra as Record<string, unknown>);
+
+export const fetchData = createAsyncThunk(
+  '/hotels',
+  async (_, thunkApi) => {
+    const { extra, rejectWithValue } = thunkApi;
+    if (!isAxiosInstance(extra)) {
+      return rejectWithValue(new Error('we expect axios'));
+    }
+    const { data } = await extra.get<Offer[]>('/hotels');
+    return data;
+  },
+);
